Litter paint refactoring
Refactored litter paint code to be in paint folder
