+++ src/openrct2/entity/Litter.cpp
@@ -155,47 +155,5 @@ namespace OpenRCT2
         stream << SubType;
         stream << creationTick;
     }
-
-    struct LitterSprite
-    {
-        uint16_t base_id;
-        uint8_t direction_mask;
-    };
-
     /** rct2: 0x0097EF6C */
-    static constexpr LitterSprite kLitterSprites[] = {
-        { SPR_LITTER_SICK, 0x1 },
-        { SPR_LITTER_SICK_ALT, 0x1 },
-        { SPR_LITTER_EMPTY_CAN, 0x1 },
-        { SPR_LITTER_RUBBISH, 0x1 },
-        { SPR_LITTER_EMPTY_BURGER_BOX, 0x1 },
-        { SPR_LITTER_EMPTY_CUP, 0x1 },
-        { SPR_LITTER_EMPTY_BOX, 0x1 },
-        { SPR_LITTER_EMPTY_BOTTLE, 0x1 },
-        { SPR_LITTER_EMPTY_BOWL_RED, 0x3 },
-        { SPR_LITTER_EMPTY_DRINK_CART, 0x3 },
-        { SPR_LITTER_EMPTY_JUICE_CUP, 0x3 },
-        { SPR_LITTER_EMPTY_BOWL_BLUE, 0x3 },
-    };
-
-    void Litter::Paint(PaintSession& session, int32_t imageDirection) const
-    {
-        PROFILED_FUNCTION();
-
-        auto& rt = session.rt;
-        if (rt.zoom_level > ZoomLevel{ 0 })
-            return; // If zoomed at all no litter drawn
-
-        // litter has no sprite direction so remove that
-        imageDirection >>= 3;
-        // Some litter types have only 1 direction so remove
-        // anything that isn't required.
-        imageDirection &= kLitterSprites[EnumValue(SubType)].direction_mask;
-
-        uint32_t image_id = imageDirection + kLitterSprites[EnumValue(SubType)].base_id;
-
-        // In the following call to PaintAddImageAsParent, we add 4 (instead of 2) to the
-        // bound_box_offset_z to make sure litter is drawn on top of railways
-        PaintAddImageAsParent(session, ImageId(image_id), { 0, 0, z }, { { -4, -4, z + 4 }, { 5, 5, -1 } });
-    }
 } // namespace OpenRCT2
